Added Game and Window classes, added splashscreen
Details: Created a game class to control the updating of the screen. Created a window class to create a window, draw to the window, close the window if the user presses the close button. A splashscreen was created which shows the name of the game as well as which buttons to press to play the game. Main was refactored to create a game object and control the game loop.

Issues: On account of the new classes, the animation of the player has no longer been implemented. A class must be created to control the opening of files, as there is a lot of repetition in the code to do this currently