Updated enemy class
Details: Recreated the enemy class by adding random starting angles to the enemies from the center of the screen. When the enemy reaches the perimeter of the circular screen, the enemy is repawned in the center of the screen again with a new random angle.

Issues: A sprite has not been used for the enemy yet. It is currently a rectangle.