Added movingEntity class, refactored the code
Details: Added a movingEntity base class from where the player and enemy classes inherit their basic movement functions. Fixed up spelling in the PlayerBullet function. Added a function to the window class to draw a movingEntity object so that the window now controls the drawing of the player and the enemy.

Issues: Unable to set a texture for the enemy. The update function in the movingEntity class has an unused input when used in the enemy class