Added basic collision testing
Details: Added collision testing between the enemy and the player, as well as between the enemy and the player bullets. Added a cleanup function to remove all "dead" objects

Issues: There is a long lag time between when the player dies and when the lose screen appears. The radius in which the collisions occur is too large