Created a satellite class, limited the number of enemies on screen at any time to 3
Details: Created a three satellites that are created in front of the player and rotate in a small circle. Enemy bullet class was refactored to allow the satellite class to use it.

Issues: Satellites shoot towards the players initial position, they do not track the player. The player's gun is not upgraded when all three are killed.