Removed dependency of Game class on SFML, added error check for loading of textures
Details: Created a clock class in order to separate the Game class (logic layer) from SFML (presentation layer). Game class is now separated from SFML

Issues: The number of lives and enemies remaining were previously displayed in the game class. The responsibility of displaying this needs to be moved to the Window class